refactor: 훅 파일 구조 개선
